refactor: streamline end-to-end tests by removing flaky health card flow
Temporarily remove healthCard flow
