+++ app/src/androidTest/java/com/github/warnastrophy/e2e/EndToEndM2Test.kt
@@ -1,12 +1,9 @@
 package com.github.warnastrophy.e2e
 
 import androidx.compose.ui.test.assertIsDisplayed
-import androidx.compose.ui.test.assertTextContains
 import androidx.compose.ui.test.onNodeWithTag
-import androidx.compose.ui.test.performClick
 import com.github.warnastrophy.core.data.repository.ContactRepositoryProvider
 import com.github.warnastrophy.core.data.repository.HealthCardRepositoryProvider
-import com.github.warnastrophy.core.ui.features.dashboard.DashboardHealthCardTestTags
 import com.github.warnastrophy.core.ui.features.dashboard.DashboardScreenTestTags
 import com.github.warnastrophy.core.ui.navigation.NavigationTestTags
 import org.junit.Before
@@ -40,42 +37,12 @@ class EndToEndM2Test : EndToEndUtils() {
     composeTestRule.onNodeWithTag(DashboardScreenTestTags.ROW_TWO_SMALL_CARDS).assertIsDisplayed()
   }
 
-  /**
-   * E2E test to verify the health card flow.
-   * - Starts on the Dashboard screen.
-   * - Clicks on the Health Card widget.
-   * - Fills and saves a new Health Card.
-   * - Navigates to the Profile screen.
-   * - Opens the Health Card from the profile.
-   * - Updates the Health Card content and saves it.
-   * - Deletes the Health Card.
-   */
   @Test
-  fun create_and_update_health_card_flow() {
+  fun create_edit_and_delete_contact() {
     setContent()
-    composeTestRule
-        .onNodeWithTag(NavigationTestTags.TOP_BAR_TITLE)
-        .assertTextContains("Dashboard", ignoreCase = true)
-
-    composeTestRule.onNodeWithTag(DashboardHealthCardTestTags.CARD).assertIsDisplayed()
-    composeTestRule.onNodeWithTag(DashboardHealthCardTestTags.CARD).performClick()
-
-    createHealthCard(fullName = "John Doe", allergies = "Peanuts")
-
-    composeTestRule.onNodeWithTag(NavigationTestTags.TAB_PROFILE).performClick()
-    composeTestRule
-        .onNodeWithTag(NavigationTestTags.TOP_BAR_TITLE)
-        .assertTextContains("Profile", ignoreCase = true)
-
-    composeTestRule.onNodeWithTag(NavigationTestTags.HEALTH_CARD).performClick()
-    composeTestRule
-        .onNodeWithTag(NavigationTestTags.TOP_BAR_TITLE)
-        .assertTextContains("Emergency Card", ignoreCase = true)
-
-    editHealthCard(fullName = "Johnathan Smith", allergies = "Peanuts, Shellfish")
-
-    composeTestRule.waitForIdle()
-
-    deleteHealthCard()
+    addNewContact()
+    editContact(saveChanges = false)
+    editContact(saveChanges = true)
+    deleteContact()
   }
 }
